(#26) Добавить механизм DB-миграций
Тип всех id изменён с int на long.

+++ backend/src/main/java/com/github/viktor235/rggassistant/models/entitys/modifiers/CollectedEffect.java
@@ -15,7 +15,12 @@ public class CollectedEffect {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private long id;
+
+//    @Schema(description = "Date the effect was added to the collection")
+//    @Column(name = "collection_date", insertable = false)
+//    @GeneratedValue(strategy = GenerationType.IDENTITY)
+//    private ZonedDateTime collectionDate;
 
     @JoinColumn(name = "effect_id", referencedColumnName = "id")
     @ManyToOne
